refactor: rename module to mapping
